extract vue as plugin; initfrntend rewrite (commpletely untested)

+++ gdaps/frontend/management/commands/initfrontend.py
@@ -2,7 +2,6 @@ import os
 import logging
 import shutil
 import subprocess
-import sys
 
 import django
 from django.conf import settings
@@ -11,12 +10,11 @@ from django.apps import apps
 from django.template import Context
 from django.utils.version import get_docs_version
 
+from gdaps import ExtensionPoint
 from gdaps.conf import gdaps_settings
+from gdaps.frontend.api import IFrontendEngines
 
 logger = logging.getLogger(__name__)
-
-# TODO: allow dynamic engines
-_engines = ["vue"]
 
 
 class Command(BaseCommand):
@@ -29,36 +27,54 @@ class Command(BaseCommand):
         (".py-tpl", ".py"),
     )
 
+    # TODO: allow dynamic engines
+    _engines = ExtensionPoint(IFrontendEngines)
+
     def is_supported_engine(self, engine):
-        if engine not in _engines:
+        engine_names = [engine.name for engine in self._engines]
+        if engine not in engine_names:
             raise CommandError(
-                f"'{engine}' is not supported as frontend engine. Available engines are: {_engines} "
+                f"'{engine.name}' is not supported as frontend engine. Available engines are: {engine_names} "
             )
 
     def handle(self, **options):
-        options["frontend_dir"] = settings.GDAPS["FRONTEND_DIR"]
         try:
-            self.engine = settings.GDAPS["FRONTEND_ENGINE"]
+            frontend_dir = settings.GDAPS["FRONTEND_DIR"]
         except:
-            self.engine = "vue"
+            frontend_dir = "frontend"
 
-        # TODO: add verbosity flag
         self.verbosity = options["verbosity"]
 
-        # TODO: allow dynamic (plugin?) engines
-        if not self.engine in ["vue"]:
+        if len(self._engines) == 0:
+            raise CommandError("There is no frontend engine available.")
+
+        self.engine = None  # type: IFrontendEngines
+        for engine in self._engines:
+            if engine.name == settings.GDAPS["FRONTEND_ENGINE"]:
+                self.engine = engine
+                break
+        else:
             raise CommandError(f"Engine [self.engine] not supported.")
 
-        # create a frontend/ directory in the Django root
-        frontend_path = os.path.join(settings.BASE_DIR, options["frontend_dir"])
-
-        if os.path.exists(frontend_path):
+        if not self.engine:
             raise CommandError(
-                "There already seems to be a frontend directory with that name. "
-                f"Please delete the '{options['frontend_dir']}' directory if you want to create a new one, "
+                "No frontend engine is selected. Please select one in settings.py using GDAPS['FRONTEND_ENGINE')"
             )
 
-        os.mkdir(frontend_path)
+        # create a frontend/ directory in the Django root
+        frontend_path = os.path.abspath(
+            os.path.expanduser(os.path.join(settings.BASE_DIR, options["frontend_dir"]))
+        )
+
+        options["files"] += self.engine.files
+
+        try:
+            os.mkdir(frontend_path)
+        except FileExistsError:
+            raise CommandError(
+                "There already seems to be a frontend directory with that name. "
+                f"Please delete the '{frontend_dir}' directory if you want to create a new one."
+            )
 
         extra_files = []
         for file in options["files"]:
@@ -72,13 +88,15 @@ class Command(BaseCommand):
         project_name = self._django_root
         project_title = self._django_root.title().replace("_", " ")
         files = []
-        extensions = []
+        extensions = ["js"]
 
         context = Context(
             {
                 **options,
                 "project_name": project_name,
                 "project_title": project_title,
+                "frontend_dir": frontend_dir,
+                "frontend_path": frontend_path,
                 "docs_version": get_docs_version(),
                 "django_version": django.__version__,
             },
@@ -101,38 +119,35 @@ class Command(BaseCommand):
 
         for root, dirs, files in os.walk(template_dir):
 
-            path_rest = root[prefix_length:]
-            # FIXME
-            relative_dir = path_rest.replace(base_name, name)
-            if relative_dir:
-                target_dir = os.path.join(frontend_path, relative_dir)
-                os.makedirs(target_dir, exist_ok=True)
-
             for dirname in dirs[:]:
                 if dirname.startswith(".") or dirname == "__pycache__":
                     dirs.remove(dirname)
 
             for filename in files:
+                # FIXME: use Js specific file endings here
                 if filename.endswith((".pyo", ".pyc", ".py.class")):
                     # Ignore some files as they cause various breakages.
                     continue
                 old_path = os.path.join(root, filename)
-                new_path = os.path.join(
-                    frontend_path, relative_dir, filename.replace(base_name, name)
-                )
+
+                new_path = os.path.join(frontend_path, filename)
                 for old_suffix, new_suffix in self.rewrite_template_suffixes:
                     if new_path.endswith(old_suffix):
                         new_path = new_path[: -len(old_suffix)] + new_suffix
                         break  # Only rewrite once
 
-                # FIXME: maybe foo.py AND foo.py-tpl exist. Print arror! Don't overwrite!
+                if os.path.exists(new_path):
+                    raise CommandError(
+                        f"{new_path} already exists, overlaying a "
+                        "frontend file into an existing directory won't replace conflicting files"
+                    )
 
                 # Only render intended files, as we don't want to
                 # accidentally render Django templates files
                 if new_path.endswith(extensions) or filename in extra_files:
                     with open(old_path, encoding="utf-8") as template_file:
                         content = template_file.read()
-                    template = Engine().from_string(content)
+                    template = django.template.Engine().from_string(content)
                     content = template.render(context)
                     with open(new_path, "w", encoding="utf-8") as new_file:
                         new_file.write(content)
@@ -140,80 +155,35 @@ class Command(BaseCommand):
                     shutil.copyfile(old_path, new_path)
 
                 if self.verbosity >= 2:
-                    self.stdout.write("Creating %s\n" % new_path)
+                    self.stdout.write(f"Creating {new_path}\n")
                 try:
                     shutil.copymode(old_path, new_path)
                     self.make_writeable(new_path)
                 except OSError:
                     self.stderr.write(
-                        "Notice: Couldn't set permission bits on %s. You're "
-                        "probably using an uncommon filesystem setup. No "
-                        "problem." % new_path,
-                        self.style.NOTICE,
+                        f"Notice: Couldn't set permission bits on {new_path}. You're "
+                        "probably using an uncommon filesystem setup. No problem."
                     )
 
-    def handle(self, *args, **options):
+        self.engine.initialize(frontend_path)
 
-        # check if the engine is supported
+        # build
+        # subprocess.check_call(
+        #     "npm run build --prefix {base_dir}/{plugin}/frontend".format(
+        #         plugin=target, base_dir=settings.BASE_DIR
+        #     ),
+        #     shell=True,
+        # )
 
-        # preparation
-        options["project_name"] = self._django_root
-        options["project_title"] = self._django_root.title().replace("_", " ")
-        options["files"] = []
-        options["extensions"] = []
+        # ask the user to be sure to copy the static files
+        # subprocess.check_call('./manage.py collectstatic'.format(plugin=target, base_dir=settings.BASE_DIR), shell=True)
 
-        # extend a plugin with vuejs
-        if options["engine"] == "vue":
-
-            if not os.path.exists(frontend_path):
-                os.mkdir(frontend_path)
-
-            # create files
-            template = os.path.join(
-                apps.get_app_config("frontend").path,
-                "management",
-                "templates",
-                "frontend",
-                "vue",
-            )
-            options["files"] += [
-                ".gitignore",
-                "babel.config.js",
-                "package.json",  # contains dependencies
-                "vue.config.js",
-                "src/App.vue",
-                "src/main.js",
-                "src/plugins.js",
-                "src/assets/logo.png",
-                "src/components/HelloWorld.vue",
-            ]
-
-            try:
-                super().handle(
-                    "app",
-                    options["project_name"] + "_frontend",
-                    target=frontend_path,
-                    template=template,
-                    **options,
-                )
-
-                # yarn install vue
-                # FIXME: check if yarn is available
-                subprocess.check_call(
-                    "yarn install --cwd {}".format(frontend_path), shell=True
-                )
-
-            except Exception as e:
-                shutil.rmtree(frontend_path)
-                raise e
-
-            # build
-            # subprocess.check_call(
-            #     "npm run build --prefix {base_dir}/{plugin}/frontend".format(
-            #         plugin=target, base_dir=settings.BASE_DIR
-            #     ),
-            #     shell=True,
-            # )
-
-            # ask the user to be sure to copy the static files
-            # subprocess.check_call('./manage.py collectstatic'.format(plugin=target, base_dir=settings.BASE_DIR), shell=True)
+    def make_writeable(self, filename):
+        """
+        Make sure that the file is writeable.
+        Useful if our source is read-only.
+        """
+        if not os.access(filename, os.W_OK):
+            st = os.stat(filename)
+            new_permissions = stat.S_IMODE(st.st_mode) | stat.S_IWUSR
+            os.chmod(filename, new_permissions)
